Check for multiple dollar signs in binbuf_restore
In the event that the first dollar sign we come across isn't related to a
dollarsym atom, we need to keep looking. Otherwise, there will be
inconsistencies between a patch's state upon saving, and its state after
closing and re-opening.

Example: in the message "$$1", the 1st $ is actually a dollarsym's
prefix, so we need to check the 2nd $ as well.